% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data function}
\usage{
get_data(
  registry,
  indicator,
  selected_orgnr,
  api = "https://prod-mong-api.skde.org/data",
  query = "indicators?context=caregiver&type=ind"
)
}
\arguments{
\item{registry}{Name of quality registry}

\item{indicator}{Indicator ID}

\item{selected_orgnr}{Which orgnr to keep data from}

\item{api}{url to api}

\item{query}{query parameters to api}
}
\description{
Get data function
}
